Backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.0.0
Pillow==10.0.0
qrcode[pil]==7.4.2
djangorestframework-simplejwt==5.3.0